requirements.txt
streamlit==1.30.0
requests==2.28.2
openai==1.55.3
azure-core==1.32.0
azure-search-documents==11.4.0b11
tiktoken==0.5.2
#chromadb==0.5.0
